server mod
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@10180 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/innbbsd/bbsnnrp.c
@@ -336,7 +336,7 @@ int main(int argc, char **argv)
         if (isfile(active)) {
             strncpy(BBSNNRP.activefile, active, sizeof BBSNNRP.activefile);
         } else if (strchr(active, '/') == NULL) {
-            sprintf(BBSNNRP.activefile, "%s/innd/%.*s", BBSHOME, sizeof BBSNNRP.activefile - 7 - strlen(BBSHOME), active);
+            sprintf(BBSNNRP.activefile, "%s/innd/%.*s", BBSHOME, (int)(sizeof BBSNNRP.activefile - 7 - strlen(BBSHOME)), active);
         } else {
             strncpy(BBSNNRP.activefile, active, sizeof BBSNNRP.activefile);
         }
